Update back icon of navigation header
